Move web-stuff to web module
